+++ spec/Akeneo/Crowdin/TranslationDirectoriesCreatorSpec.php
@@ -5,8 +5,8 @@ namespace spec\Akeneo\Crowdin;
 use Akeneo\Crowdin\Api\AddDirectory;
 use Akeneo\Crowdin\Client;
 use Akeneo\Crowdin\TranslationProjectInfo;
-use Akeneo\System\TargetResolver;
-use Akeneo\System\TranslationFile;
+use Akeneo\Nelson\TargetResolver;
+use Akeneo\Nelson\TranslationFile;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Psr\Log\LoggerInterface;
