+++ src/Akeneo/Nelson/TranslationFilesCleaner.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Akeneo\System;
+namespace Akeneo\Nelson;
 
+use Akeneo\System\Executor;
 use SplFileInfo;
 use Symfony\Component\Finder\Finder;
