+++ src/Akeneo/Resources/config/crowdin.yml
@@ -27,21 +27,21 @@ services:
         arguments:
             - '@crowdin.client'
             - '@event_dispatcher'
-            - '@akeneo.system.target_resolver'
+            - '@nelson.target_resolver'
 
     crowdin.translation_files.directories_creator:
         class: Akeneo\Crowdin\TranslationDirectoriesCreator
         arguments:
             - '@crowdin.client'
             - '@event_dispatcher'
-            - '@akeneo.system.target_resolver'
+            - '@nelson.target_resolver'
 
     crowdin.translation_files.files_creator:
         class: Akeneo\Crowdin\TranslationFilesCreator
         arguments:
             - '@crowdin.client'
             - '@event_dispatcher'
-            - '@akeneo.system.target_resolver'
+            - '@nelson.target_resolver'
 
     crowdin.translation_files.project_info:
         class: Akeneo\Crowdin\TranslationProjectInfo
